app/api/pyproject.toml
[project]
name = "mytravel-api"
version = "1.0.0"
description = "MyTravel FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pydantic[email]>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "anthropic>=0.30",
    "httpx>=0.27",
    "resend>=2.0",
    "sentry-sdk[fastapi]>=2.0",
    "authlib>=1.3",
    "itsdangerous>=2.2",
    "python-multipart>=0.0.9",
    "email-validator>=2.2",
    "mako>=1.3.10",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
